# wealth-tech-openapi.yaml
openapi: 3.1.0
info:
  title: Nevis Search Service API
  version: 1.1.0
  description: |
    Search API for clients and their documents.

    - Clients are searched via keyword matching (email, name, description)
    - Documents are searched via semantic similarity (embeddings)
    - Results are returned in separate sections because scores are not directly comparable

servers:
  - url: http://localhost:8080

paths:
  /clients/{id}:
    get:
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Client found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found

  /documents/{id}:
    get:
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Document not found

  /clients:
    post:
      summary: Create client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  /clients/{clientId}/documents/{documentId}:
    get:
      summary: Get document by client ID and document ID
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Document found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Client or document not found

  /clients/{id}/documents:
    post:
      summary: Create document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /search:
    get:
      summary: Search clients and documents
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 127
        - name: clientLimit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 5
            maximum: 20
        - name: documentLimit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
            maximum: 50
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"

components:
  schemas:

    CreateClientRequest:
      type: object
      required: [first_name, last_name, email]
      properties:
        first_name:
          type: string
          minLength: 1
        last_name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        description:
          type: string
        social_links:
          type: array
          items:
            type: object
            required: [url]
            properties:
              url:
                type: string
                minLength: 1

    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        description:
          type: string
        social_links:
          type: array
          items:
            type: string

    CreateDocumentRequest:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1

    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        summary:
          type: string
        created_at:
          type: string
          format: date-time

    SearchResponse:
      type: object
      properties:
        query:
          type: string
        clients:
          type: array
          items:
            $ref: "#/components/schemas/ClientSearchResult"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSearchResult"
        errors:
          type: array
          description: Partial search errors. Empty when all search sources are available.
          items:
            type: string

    ClientSearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        score:
          type: number
          format: float

    DocumentSearchResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        score:
          type: number
          format: float
        title:
          type: string
        summary:
          type: string
        matched_text:
          type: string